+++ src/main/java/de/markusbordihn/playercompanions/client/renderer/companions/LizardRenderer.java
@@ -19,6 +19,10 @@
 
 package de.markusbordihn.playercompanions.client.renderer.companions;
 
+import java.util.EnumMap;
+import java.util.Map;
+
+import net.minecraft.Util;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.MobRenderer;
 import net.minecraft.resources.ResourceLocation;
@@ -26,19 +30,35 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 
+import de.markusbordihn.playercompanions.Constants;
 import de.markusbordihn.playercompanions.client.model.LizardModel;
 import de.markusbordihn.playercompanions.client.renderer.ClientRenderer;
+import de.markusbordihn.playercompanions.entity.PlayerCompanionVariant;
 import de.markusbordihn.playercompanions.entity.companions.Lizard;
 
 @OnlyIn(Dist.CLIENT)
 public class LizardRenderer extends MobRenderer<Lizard, LizardModel<Lizard>> {
 
+  // Variant Textures
+  protected static final Map<PlayerCompanionVariant, ResourceLocation> TEXTURE_BY_VARIANT =
+      Util.make(new EnumMap<>(PlayerCompanionVariant.class), hashMap -> {
+        hashMap.put(PlayerCompanionVariant.DEFAULT,
+            new ResourceLocation(Constants.MOD_ID, "textures/entity/lizard/lizard_default.png"));
+        hashMap.put(PlayerCompanionVariant.DESERT,
+            new ResourceLocation(Constants.MOD_ID, "textures/entity/lizard/lizard_desert.png"));
+        hashMap.put(PlayerCompanionVariant.GREEN,
+            new ResourceLocation(Constants.MOD_ID, "textures/entity/lizard/lizard_green.png"));
+      });
+  protected static final ResourceLocation DEFAULT_TEXTURE =
+      TEXTURE_BY_VARIANT.get(PlayerCompanionVariant.DEFAULT);
+
   public LizardRenderer(EntityRendererProvider.Context context) {
     super(context, new LizardModel<>(context.bakeLayer(ClientRenderer.LIZARD)), 0.5F);
   }
 
+  @Override
   public ResourceLocation getTextureLocation(Lizard entity) {
-    return entity.getTextureLocation();
+    return TEXTURE_BY_VARIANT.getOrDefault(entity.getVariant(), DEFAULT_TEXTURE);
   }
 
 }