+++ src/main/java/de/markusbordihn/playercompanions/client/renderer/companions/RaptorRenderer.java
@@ -19,21 +19,43 @@
 
 package de.markusbordihn.playercompanions.client.renderer.companions;
 
+import java.util.EnumMap;
+import java.util.Map;
+
+import net.minecraft.Util;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.MobRenderer;
 import net.minecraft.resources.ResourceLocation;
 
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+
+import de.markusbordihn.playercompanions.Constants;
 import de.markusbordihn.playercompanions.client.model.RaptorModel;
 import de.markusbordihn.playercompanions.client.renderer.ClientRenderer;
+import de.markusbordihn.playercompanions.entity.PlayerCompanionVariant;
 import de.markusbordihn.playercompanions.entity.companions.Raptor;
 
+@OnlyIn(Dist.CLIENT)
 public class RaptorRenderer extends MobRenderer<Raptor, RaptorModel<Raptor>> {
+
+  // Variant Textures
+  protected static final Map<PlayerCompanionVariant, ResourceLocation> TEXTURE_BY_VARIANT =
+      Util.make(new EnumMap<>(PlayerCompanionVariant.class), hashMap -> {
+        hashMap.put(PlayerCompanionVariant.DEFAULT,
+            new ResourceLocation(Constants.MOD_ID, "textures/entity/raptor/raptor_default.png"));
+        hashMap.put(PlayerCompanionVariant.PINK,
+            new ResourceLocation(Constants.MOD_ID, "textures/entity/raptor/raptor_pink.png"));
+      });
+  protected static final ResourceLocation DEFAULT_TEXTURE =
+      TEXTURE_BY_VARIANT.get(PlayerCompanionVariant.DEFAULT);
 
   public RaptorRenderer(EntityRendererProvider.Context context) {
     super(context, new RaptorModel<>(context.bakeLayer(ClientRenderer.RAPTOR)), 0.7F);
   }
 
+  @Override
   public ResourceLocation getTextureLocation(Raptor entity) {
-    return entity.getTextureLocation();
+    return TEXTURE_BY_VARIANT.getOrDefault(entity.getVariant(), DEFAULT_TEXTURE);
   }
 }