+++ src/main/java/de/markusbordihn/playercompanions/entity/type/collector/CollectorEntityFloating.java
@@ -22,7 +22,7 @@ package de.markusbordihn.playercompanions.entity.type.collector;
 import java.util.EnumSet;
 import java.util.Map;
 import java.util.Set;
-import net.minecraft.resources.ResourceLocation;
+
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -43,9 +43,8 @@ public class CollectorEntityFloating extends PlayerCompanionEntityFloating {
       EnumSet.of(AggressionLevel.PASSIVE_FLEE, AggressionLevel.PASSIVE);
 
   public CollectorEntityFloating(EntityType<? extends PlayerCompanionEntity> entityType,
-      Level level, Map<PlayerCompanionVariant, ResourceLocation> textureByVariant,
-      Map<PlayerCompanionVariant, Item> companionItemByVariant) {
-    super(entityType, level, textureByVariant, companionItemByVariant);
+      Level level, Map<PlayerCompanionVariant, Item> companionItemByVariant) {
+    super(entityType, level, companionItemByVariant);
     this.setAggressionLevel(getDefaultAggressionLevel());
 
     // Shared Collector Features